Create wayback module
